+++ app/src/core/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './providerTypes'
 export * from './providerStore'
+export * from './providerManager'